requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
transformers==4.35.0
torch==2.1.0
Pillow==10.0.1
PyPDF2==3.0.1
pytesseract==0.3.10
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
